Guard engine arithmetic against overflow crashes
Three user-reachable crashes (each escaped the engine and crashed the client):
- random(min, MAXINT): b+1 overflowed to Int.MIN, nextInt threw. Compute the
  inclusive bound in Long.
- oversized int literal in calc/expressions: toInt() threw NumberFormatException
  in the lexer. Saturate to Int.MAX (value model is 32-bit).
- for(#i, .., MAXINT): counter wrapped past Int.MAX so current<=end stayed true,
  running ~1M times before the step-cap threw. Advance in Long, terminate when
  the counter leaves Int range.

Adds passing regression tests for each (193 engine tests, 0 fail).

Goal: 026-hardening-r2 (deliverable 2)

+++ engine/src/main/kotlin/dev/macromod/engine/action/builtin/StringMathActions.kt
@@ -24,7 +24,9 @@ object RandomAction : ScriptAction("random") {
             val hi = ctx.evaluate(args[1]).asInt()
             val a = minOf(lo, hi)
             val b = maxOf(lo, hi)
-            ReturnValue.of(Random.nextInt(a, b + 1))
+            // Compute the inclusive upper bound in Long: `b + 1` overflows to Int.MIN_VALUE when
+            // b == Int.MAX_VALUE, which would make nextInt(a, bound) throw (bound <= origin).
+            ReturnValue.of(Random.nextLong(a.toLong(), b.toLong() + 1).toInt())
         }
         args.size == 1 -> {
             val n = ctx.evaluate(args[0]).asInt()
